Don't declare something a long and pass it as a size_t
The two are not the same on Win64, causing type intersection to sometimes return incorrect answers (not fun!).
(cherry picked from commit b0e19c87380b88a7a907ab6227bcc4fdeb5bf3ab)
ref #14223

+++ src/jltypes.c
@@ -833,7 +833,7 @@ static jl_value_t *jl_type_intersect(jl_value_t *a, jl_value_t *b,
     if (b == (jl_value_t*)jl_any_type || b == jl_ANY_flag) return a;
     // tuple
     if (jl_is_tuple_type(a)) {
-        long alen = (long)jl_nparams(a);
+        size_t alen = jl_nparams(a);
         jl_value_t *temp=NULL;
         JL_GC_PUSH2(&b, &temp);
         if (jl_is_ntuple_type(b)) {
